+++ libs/cdk/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ethlete/cdk
 
+## 4.38.2
+
+### Patch Changes
+
+- [#2168](https://github.com/ethlete-io/ethdk/pull/2168) [`3af9d56`](https://github.com/ethlete-io/ethdk/commit/3af9d560c7ab87cdb3ae6f7f4e18c75da35440ec) Thanks [@baltruschat](https://github.com/baltruschat)! - Fix: add missing color input export
+
+## 4.38.1
+
+### Patch Changes
+
+- [#2166](https://github.com/ethlete-io/ethdk/pull/2166) [`e48eb0f`](https://github.com/ethlete-io/ethdk/commit/e48eb0fc82aecbd1360e9da451bde1c6dd89535f) Thanks [@baltruschat](https://github.com/baltruschat)! - Add color input
+
+## 4.38.0
+
+### Minor Changes
+
+- [#2152](https://github.com/ethlete-io/ethdk/pull/2152) [`bd090e6`](https://github.com/ethlete-io/ethdk/commit/bd090e65ecd95def25be60562238443abd1ffa06) Thanks [@Gw3i](https://github.com/Gw3i)! - Add `formChangesTracker` util to check for changes, restore, refresh or get the default value of a form or form control
+
+## 4.37.2
+
+### Patch Changes
+
+- [`2df2e5a`](https://github.com/ethlete-io/ethdk/commit/2df2e5a3d6ed600ddc6bfb8533675bfa14d314d4) Thanks [@TomTomB](https://github.com/TomTomB)! - Fix sortable state not updating if the bound form control value changes
+
 ## 4.37.1
 
 ### Patch Changes